third_party/blink/web_tests/external/wpt/mediacapture-depth/dictionary-manual.https-expected.txt

This is a testharness.js-based test.
PASS MediaTrackSupportedConstraints dictionary include attributes are correct
FAIL MediaTrackCapabilities dictionary of depth include attributes are correct assert_equals: Expect dictionary.videoKind to be string expected "string" but got "undefined"
PASS MediaTrackConstraintSet dictionary of depth include attributes are correct
PASS MediaTrackSettings dictionary of depth include attributes are correct
FAIL MediaTrackCapabilities dictionary of color include attributes are correct assert_equals: Expect dictionary.videoKind to be string expected "string" but got "undefined"
PASS MediaTrackConstraintSet dictionary of color include attributes are correct
PASS MediaTrackSettings dictionary of color include attributes are correct
Harness: the test ran to completion.
